feat: move batch job util closer to the batch job component
